Support per-method `Send` bounds.
This allows the trait designer to require that the futures
implementations use for the async methods, implement Send. Unfortunately
however, it is currently nontrivial to require Send if and only if the
implementor type itself is send, which perhaps would even require
another language feature such as `async Send`.

I am also planning on adding a convenience top-level attribute to
require all methods to be Send.